refactor(bot): PR-4 — slack/MM app modules + Cursor client factory
- Move Bolt/WebSocket wiring to adapters/*/app.py; thin module launchers
- build_cursor_client in infra/cursor_client_factory.py
- docs + ruff F401 for launcher re-exports; add stacked PR workflow note
