+++ docs/stacked-pr-workflow.md
@@ -1,0 +1,10 @@
+# 積み上げ PR（スタック）でのボット改修
+
+大きなリファクタでは **レビュー単位が追いにくくなる誤り** が起きやすい。エージェントやローカル開発でも次を守る。
+
+1. **事前に PR 境界を決める**（例: テストだけ → ドメイン/Repo → ChatAdapter → 並行性 → ランチャー分割）。境界ごとに「この PR で終わる変更」を一文で書く。
+2. **境界ごとにブランチまたはコミットで止める**。スナップショット全体を一気に当てない。途中状態でも `pytest` / `ruff` が緑であることを確認してから次へ。
+3. **「後で直す」モジュール間の不整合を作らない**。たとえば `run_cursor_reply` の引数と呼び出し側は同一 PR で揃える。
+4. **レビュー依頼はスタックの先頭から**。依存関係が複雑なときは `git range-diff` や GitHub の stacked PR / ベースブランチの明示で読み手の負荷を下げる。
+
+詳細なボット構成は [bot.md](bot.md) と [architecture.md](architecture.md) を参照。
