+++ src/bot/infra/__init__.py
@@ -1,0 +1,1 @@
+"""Infrastructure helpers (factories for external clients)."""
